Refactor Grave Traffic Model to new format
