+++ c_src/include/PrepareSim.h
@@ -8,15 +8,14 @@
 #include <algorithm>
 #include <time.h>
 #include <stdlib.h>
+#include <filesystem>
 #include "ConfigData.h"
-#include "TrafficFiles.h"
 #include "Lane.h"
 #include "LaneGenTraffic.h"
 #include "LaneFileTraffic.h"
 #include "VehicleGenerator.h"
 #include "VehicleTrafficFile.h"
 #include "VehicleBuffer.h"
-#include "TrafficData.h"
 #include "ReadILFile.h"
 #include "BridgeFile.h"
 #include "Bridge.h"
