+++ c_src/src/VehModelDataGrave.cpp
@@ -1,11 +1,14 @@
 #include "VehModelDataGrave.h"
 #include "ConfigData.h"
 
-//extern CConfigData g_ConfigData;
 
 CVehModelDataGrave::CVehModelDataGrave(CVehicleClassification_sp pVC, CLaneFlowComposition lfc)
 	: CVehicleModelData(eVM_Grave, pVC, lfc, 5) // MAGIC NUMBER - truck class count of the Grave Model
 {
+	GVWRange range;
+	m_v4AxleWeight.assign(12, range);
+	m_v5AxleWeight.assign(12, range);
+
 	ReadDataIn();
 }
 
@@ -14,33 +17,162 @@ CVehModelDataGrave::~CVehModelDataGrave()
 
 }
 
+////////// THE GETS ////////////////
+
 std::vector<double>	CVehModelDataGrave::GetGVWRange(size_t iTruck, size_t iRange)
 {
-	return m_TD.GetGVWRange(iTruck, iRange);
-}
+	std::vector<double> data;
+	GVWRange range;
+	if(iRange > 11)
+		iRange = 11;
+	if( iTruck == 4)
+		range = m_v4AxleWeight[iRange];
+	else
+		range = m_v5AxleWeight[iRange];
 
+	data.push_back(	range.W1.Mean );
+	data.push_back(	range.W2.Mean );
+	data.push_back(	range.WT.Mean );
+	data.push_back(	range.W1.StdDev );
+	data.push_back(	range.W2.StdDev );
+	data.push_back(	range.WT.StdDev );
+
+	return data;
+}
 
 CTriModalNormal	CVehModelDataGrave::GetSpacingDist(size_t iTruck, size_t iSpace)
 {
-	return m_TD.GetSpacingDist(iTruck, iSpace);
+	switch( iTruck )
+	{
+		case 2:
+			return m_v2AxleSpacings[iSpace];
+		case 3:
+			return m_v3AxleSpacings[iSpace];
+		case 4:
+			return m_v4AxleSpacings[iSpace];
+		default:
+			return m_v5AxleSpacings[iSpace];
+	}
 }
-
 
 CTriModalNormal	CVehModelDataGrave::GetAxleWeightDist(size_t iTruck, size_t iAxle)
 {
-	return m_TD.GetAxleWeightDist(iTruck, iAxle);
+	if( iTruck == 2 )
+		return m_v2AxleWeight[iAxle];
+	else
+		return m_v3AxleWeight[iAxle];
 }
-
 
 CTriModalNormal	CVehModelDataGrave::GetTrackWidthDist(size_t iTruck, size_t iAxle)
 {
-	return m_TD.GetTrackWidthDist(iTruck, iAxle);
+	switch( iTruck )
+	{
+		case 2:
+			return m_v2AxleTrackWidth[iAxle];
+		case 3:
+			return m_v3AxleTrackWidth[iAxle];
+		case 4:
+			return m_v4AxleTrackWidth[iAxle];
+		default:
+			return m_v5AxleTrackWidth[iAxle];
+	}
 }
-
 
 CTriModalNormal	CVehModelDataGrave::GetGVW(size_t dir, size_t iTruck)
 {
-	return m_TD.GetGVW(dir, iTruck);
+	if(dir == 1)
+		return m_vDir1GVW[iTruck - 2];
+	else
+		return m_vDir2GVW[iTruck - 2];
+}
+
+CTriModalNormal CVehModelDataGrave::GetSpeed(std::size_t dir)
+{
+	if (dir == 1)
+		return m_vSpeed[0];
+	else
+		return m_vSpeed[1];
+}
+
+/////////////// THE SETS //////////////////
+
+void CVehModelDataGrave::Add2AxleSpacings(std::vector<CTriModalNormal> vSpace)
+{
+	m_v2AxleSpacings = vSpace;
+}
+
+void CVehModelDataGrave::Add3AxleSpacings(std::vector<CTriModalNormal> vSpace)
+{
+	m_v3AxleSpacings = vSpace;
+}
+
+void CVehModelDataGrave::Add4AxleSpacings(std::vector<CTriModalNormal> vSpace)
+{
+	m_v4AxleSpacings = vSpace;
+}
+
+void CVehModelDataGrave::Add5AxleSpacings(std::vector<CTriModalNormal> vSpace)
+{
+	m_v5AxleSpacings = vSpace;
+}
+
+void CVehModelDataGrave::Add2AxleTrackWidth(std::vector<CTriModalNormal> vTrack)
+{
+	m_v2AxleTrackWidth = vTrack;
+}
+
+void CVehModelDataGrave::Add3AxleTrackWidth(std::vector<CTriModalNormal> vTrack)
+{
+	m_v3AxleTrackWidth = vTrack;
+}
+
+void CVehModelDataGrave::Add4AxleTrackWidth(std::vector<CTriModalNormal> vTrack)
+{
+	m_v4AxleTrackWidth = vTrack;
+}
+
+void CVehModelDataGrave::Add5AxleTrackWidth(std::vector<CTriModalNormal> vTrack)
+{
+	m_v5AxleTrackWidth = vTrack;
+}
+
+void CVehModelDataGrave::Add2AxleWeight(std::vector<CTriModalNormal> vAxle)
+{
+	m_v2AxleWeight = vAxle;
+}
+
+void CVehModelDataGrave::Add3AxleWeight(std::vector<CTriModalNormal> vAxle)
+{
+	m_v3AxleWeight = vAxle;
+}
+
+void CVehModelDataGrave::Add45AxleWeight(std::vector<double> data, std::size_t iTruck, std::size_t iRange)
+{
+	GVWRange range;
+	range.W1.Mean = data[0];
+	range.W2.Mean = data[1];
+	range.WT.Mean = data[2];
+	range.W1.StdDev = data[3];
+	range.W2.StdDev = data[4];
+	range.WT.StdDev = data[5];
+
+	if( iTruck == 4)
+		m_v4AxleWeight[iRange] = range;
+	else
+		m_v5AxleWeight[iRange] = range;
+}
+
+void CVehModelDataGrave::AddGVW(int dir, std::vector<CTriModalNormal> vGVW)
+{
+	if(dir == 1)
+		m_vDir1GVW = vGVW;
+	else
+		m_vDir2GVW = vGVW;
+}
+
+void CVehModelDataGrave::AddSpeed(std::vector<CTriModalNormal> vSpeed)
+{
+	m_vSpeed = vSpeed;
 }
 
 void CVehModelDataGrave::ReadDataIn()
@@ -96,7 +228,12 @@ vNHM[see paper/manual]
 void CVehModelDataGrave::ReadFile_AW23()
 {
 	filesystem::path file = m_Path / "AW2&3.csv";
-	m_CSV.OpenFile(file.string(), ",");
+	if( !m_CSV.OpenFile(file.string(), ",") )
+	{
+		std::cerr << "***ERROR: " << std::filesystem::weakly_canonical(file) << " file could not be opened" << endl;
+		system("PAUSE");
+		exit( 1 );
+	}
 
 	for (int iTruck = 2; iTruck <= 3; iTruck++)	// for each truck type
 	{
@@ -119,11 +256,11 @@ void CVehModelDataGrave::ReadFile_AW23()
 		}
 		if (iTruck == 2)
 		{
-			m_TD.Add2AxleWeight(vAxle);	// and add to data store
+			Add2AxleWeight(vAxle);	// and add to data store
 			m_CSV.getline(line);		// skip blank line
 		}
 		else
-			m_TD.Add3AxleWeight(vAxle);	// add to data store
+			Add3AxleWeight(vAxle);	// add to data store
 	}
 
 	m_CSV.CloseFile();			// and close file
@@ -132,7 +269,12 @@ void CVehModelDataGrave::ReadFile_AW23()
 void CVehModelDataGrave::ReadFile_AW45()
 {
 	filesystem::path file = m_Path / "AW4&5.csv";
-	m_CSV.OpenFile(file.string(), ",");
+	if( !m_CSV.OpenFile(file.string(), ",") )
+	{
+		std::cerr << "***ERROR: " << std::filesystem::weakly_canonical(file) << " file could not be opened" << endl;
+		system("PAUSE");
+		exit( 1 );
+	}
 
 	for (int iTruck = 4; iTruck <= 5; iTruck++)	// for each truck type
 	{
@@ -144,7 +286,7 @@ void CVehModelDataGrave::ReadFile_AW45()
 			m_CSV.getline(line);
 			for (int j = 0; j < 6; j++)			// extract 6 pieces of data
 				vData[j] = m_CSV.stringToDouble(m_CSV.getfield(j));
-			m_TD.Add45AxleWeight(vData, iTruck, iRange);
+			Add45AxleWeight(vData, iTruck, iRange);
 		}
 
 		if (iTruck == 4)
@@ -158,7 +300,12 @@ void CVehModelDataGrave::ReadFile_AW45()
 void CVehModelDataGrave::ReadFile_AS()
 {
 	filesystem::path  file = m_Path / "ASall.csv";
-	m_CSV.OpenFile(file.string(), ",");
+	if( !m_CSV.OpenFile(file.string(), ",") )
+	{
+		std::cerr << "***ERROR: " << std::filesystem::weakly_canonical(file) << " file could not be opened" << endl;
+		system("PAUSE");
+		exit( 1 );
+	}
 
 	for (int iTruck = 2; iTruck <= 5; iTruck++)	// for each truck type
 	{
@@ -183,16 +330,16 @@ void CVehModelDataGrave::ReadFile_AS()
 		switch (iTruck)
 		{
 		case 2:
-			m_TD.Add2AxleSpacings(vSpace);	// and add to data store
+			Add2AxleSpacings(vSpace);	// and add to data store
 			break;
 		case 3:
-			m_TD.Add3AxleSpacings(vSpace);	// and add to data store
+			Add3AxleSpacings(vSpace);	// and add to data store
 			break;
 		case 4:
-			m_TD.Add4AxleSpacings(vSpace);	// and add to data store
+			Add4AxleSpacings(vSpace);	// and add to data store
 			break;
 		case 5:
-			m_TD.Add5AxleSpacings(vSpace);	// and add to data store
+			Add5AxleSpacings(vSpace);	// and add to data store
 			break;
 		}
 
@@ -208,14 +355,16 @@ void CVehModelDataGrave::ReadFile_ATW()
 	filesystem::path  file = m_Path / "ATW.csv";
 	if (!m_CSV.OpenFile(file.string(), ","))
 	{
-		std::cout << "*** Warning: Axle track width file not found, using default values" << std::endl;
+		std::cout << "*** Warning: Axle track width file not found: " 
+				  << std::filesystem::weakly_canonical(file) 
+				  << " using default values" << std::endl;
 		CTriModalNormal tmn;
 		tmn.AddMode(1, CConfigData::get().Gen.TRUCK_TRACK_WIDTH, 0);	// deterministic width
 		vector<CTriModalNormal> vTrack(2, tmn); // start off with 2 axles
-		m_TD.Add2AxleTrackWidth(vTrack); // add it
-		vTrack.push_back(tmn);	m_TD.Add3AxleTrackWidth(vTrack); // add another axle and store it
-		vTrack.push_back(tmn);	m_TD.Add4AxleTrackWidth(vTrack); // add another axle and store it
-		vTrack.push_back(tmn);	m_TD.Add5AxleTrackWidth(vTrack); // add another axle and store it
+		Add2AxleTrackWidth(vTrack); // add it
+		vTrack.push_back(tmn);	Add3AxleTrackWidth(vTrack); // add another axle and store it
+		vTrack.push_back(tmn);	Add4AxleTrackWidth(vTrack); // add another axle and store it
+		vTrack.push_back(tmn);	Add5AxleTrackWidth(vTrack); // add another axle and store it
 		return;	// end the function here
 	}
 
@@ -243,16 +392,16 @@ void CVehModelDataGrave::ReadFile_ATW()
 		switch (iTruck)
 		{
 		case 2:
-			m_TD.Add2AxleTrackWidth(vTrack);	// and add to data store
+			Add2AxleTrackWidth(vTrack);	// and add to data store
 			break;
 		case 3:
-			m_TD.Add3AxleTrackWidth(vTrack);	// and add to data store
+			Add3AxleTrackWidth(vTrack);	// and add to data store
 			break;
 		case 4:
-			m_TD.Add4AxleTrackWidth(vTrack);	// and add to data store
+			Add4AxleTrackWidth(vTrack);	// and add to data store
 			break;
 		case 5:
-			m_TD.Add5AxleTrackWidth(vTrack);	// and add to data store
+			Add5AxleTrackWidth(vTrack);	// and add to data store
 			break;
 		}
 
@@ -266,7 +415,12 @@ void CVehModelDataGrave::ReadFile_ATW()
 void CVehModelDataGrave::ReadFile_GVW()
 {
 	filesystem::path  file = m_Path / "GVWpdf.csv";
-	m_CSV.OpenFile(file.string(), ",");
+	if( !m_CSV.OpenFile(file.string(), ",") )
+	{
+		std::cerr << "***ERROR: " << std::filesystem::weakly_canonical(file) << " file could not be opened" << endl;
+		system("PAUSE");
+		exit( 1 );
+	}
 
 	string line;
 	CTriModalNormal temp;
@@ -292,11 +446,11 @@ void CVehModelDataGrave::ReadFile_GVW()
 					vGVW[j - 1].AddMode(param[0], param[1], param[2]);		// GVW
 			}
 		}
-		m_TD.AddGVW(dir, vGVW);
+		AddGVW(dir, vGVW);
 		if (dir == 1)
 			m_CSV.getline(line);
 	}
 
-	m_TD.AddSpeed(vSpeed);
+	AddSpeed(vSpeed);
 	m_CSV.CloseFile();
 }
